pyproject.toml
[project]
name = "august-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.7"
dependencies = [
    "elasticsearch>=9.0.2",
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "python-dotenv>=1.1.1",
    "uvicorn==0.35.0",
    "mem0ai>=0.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "aiomysql>=0.2.0",
    "colorama>=0.4.6",
    "opentelemetry-distro[otlp]==0.55b1",
    "opentelemetry-exporter-otlp==1.34.1",
    "aiohttp>=3.12.14",
]
